mcp-server/requirements.txt
docker
kubernetes
prometheus-api-client
openai
slack-sdk
python-dotenv
flask
python-json-logger